requirements.txt
anthropic>=0.39.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
websockets>=13.0
playwright>=1.40.0
pyyaml>=6.0
edge-tts>=6.1.0
gTTS>=2.5.0
pyautogui>=0.9.54

# RAG Vector Memory (optional but recommended)
chromadb>=0.5.0
sentence-transformers>=3.0.0

# System Tray
pystray>=0.19.0
Pillow>=10.0.0
